Feat: add a pre verified create account flow to support fifthtry.com
fifthtry.com has a special signup flow where it used a `code` to create
pre-verified accounts (users don't get a verification email on their
account). This might be useful for other websites that use this package
too so we include that action here as well.

The action is not used anywhere in this package. Users of this package
should decide how/if they want to use it.